+++ .github/workflows/release.yml
@@ -1,4 +1,4 @@
-name: Create Release
+﻿name: Create Release
 
 on:
   push:
@@ -112,3 +112,4 @@ jobs:
           else
             echo "⏳ Docker image will be available shortly"
           fi
+
