+++ docker-compose.yml
@@ -60,3 +60,4 @@ networks:
   remnawave-network:
     name: remnawave-network
     driver: bridge
+    external: true
