+++ modules/handlers/__init__.py
@@ -1,0 +1,11 @@
+from . import core, users, nodes, hosts, inbounds, bulk, stats
+
+__all__ = [
+    "core",
+    "users",
+    "nodes",
+    "hosts",
+    "inbounds",
+    "bulk",
+    "stats",
+]
