+++ modules/handlers/stats/__init__.py
@@ -1,0 +1,1 @@
+from .handlers import *  # noqa: F401,F403
